Use fontconfig pattern to locate demos/markup.c used fonts
The previous implementation assumed that the japanese text line uses a
font called `Droid Sans Japanese`. However this didn't select the
correct `Droid Sans` variant installed on a Linux system, which leads to
missing glyphs.

Fontconfig allows to describe the `[Font Properties][]` with the `[Font Name][]`
format.  When adding the `lang` property Fontconfig now properly selects
the correct font and allows to render the text completely.

[Font Properties]: https://www.freedesktop.org/software/fontconfig/fontconfig-user.html#AEN21
[Font Name]: https://www.freedesktop.org/software/fontconfig/fontconfig-user.html#AEN36